[Back]🛠️fix : RecordTotalResponse 타입 변경

+++ backend/src/main/java/com/ssafy/rideus/dto/record/response/RecordTotalResponse.java
@@ -7,10 +7,10 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class RecordTotalResponse {
-    private Long total_time;
+    private Double total_time;
     private Double total_distance;
 
-    public RecordTotalResponse(Long total_time, Double total_distance) {
+    public RecordTotalResponse(Double total_time, Double total_distance) {
         this.total_time = total_time;
         this.total_distance = total_distance;
     }
